ts 파일 분리
