+++ src/apis/searchMovies.ts
@@ -2,17 +2,6 @@ import { BASE_URL } from "../constants";
 import type { MovieAPIResponse } from "../types/movie";
 
 const API_KEY = process.env.API_KEY;
-export const getPopularMovies = async (page: number = 1) => {
-    const response = await fetch(`${BASE_URL}movie/popular?api_key=${API_KEY}&language=ko-KR&page=${page}`);
-
-    if (response.ok) {
-        const data: MovieAPIResponse = await response?.json();
-        return data;
-    }
-    else {
-        throw new Error();
-    }
-};
 
 export const searchMovies = async (query: string, page: number = 1) => {
     const response = await fetch(`${BASE_URL}search/movie?api_key=${API_KEY}&query=${query}&language=ko-KR&page=${page}`);
@@ -24,4 +13,6 @@ export const searchMovies = async (query: string, page: number = 1) => {
     else {
         throw new Error();
     }
-}
+};
+
+export default searchMovies;
